# docs/swagger.yaml
basePath: /
definitions:
  server.swaggerAcceptInviteRequest:
    properties:
      password:
        type: string
    type: object
...
  server.swaggerCreateDiagnosisRequest:
    properties:
      appointment_id:
        type: string
      details:
        type: string
      diagnosis:
        type: string
      hospital_id:
        type: string
      patient_id:
        type: string
    type: object
...
  server.swaggerDiagnosisResponse:
    properties:
      appointment_id:
        type: string
      created_at:
        type: string
      details:
        type: string
      diagnosis:
        type: string
      dismissed:
        type: boolean
      doctor_id:
        type: string
      hospital_id:
        type: string
      id:
        type: string
      patient_id:
        type: string
      updated_at:
        type: string
    type: object
...
  server.swaggerInviteUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      hospital_id:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
...
  server.swaggerResetPasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
  server.swaggerSessionResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/server.swaggerUserResponse'
    type: object
  server.swaggerSignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  server.swaggerSignUpRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      hospital_id:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
...
  server.swaggerUserResponse:
    properties:
      accepted:
        type: boolean
      email:
        type: string
      first_name:
        type: string
      hospital_id:
        type: string
      id:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@clinicwise.io
    name: Clinic Wise Support
  description: REST API for the Clinic Wise healthcare management platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Clinic Wise API
  version: "1.0"
paths:
...
  /auth/invite:
    post:
      consumes:
      - application/json
      description: Sends an invitation to a non-patient user (doctor, pharmacist,
        admin). Requires authentication.
      parameters:
      - description: Invite payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerInviteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Invite a user
      tags:
      - Auth
  /auth/invites/{id}/accept:
    post:
      consumes:
      - application/json
      description: Accepts a pending invite by setting the user's password, activating
        the account.
      parameters:
      - description: Invite ID (ULID)
        in: path
        name: id
        required: true
        type: string
      - description: Accept invite payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerAcceptInviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Accept an invitation
      tags:
      - Auth
  /auth/password:
    patch:
      consumes:
      - application/json
      description: Changes the authenticated user's password.
      parameters:
      - description: Reset password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reset password
      tags:
      - Auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a session with access/refresh
        tokens.
      parameters:
      - description: Sign-in payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerSignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign in
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Registers a new patient account and returns a session with access/refresh
        tokens.
      parameters:
      - description: Sign-up payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerSignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign up a new patient
      tags:
      - Auth
  /diagnoses:
    get:
      description: Returns all diagnoses associated with the authenticated user. Requires
        doctor or patient role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.swaggerDiagnosisResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List diagnoses for the current user
      tags:
      - Diagnoses
    post:
      consumes:
      - application/json
      description: Creates a new diagnosis for a patient. Requires doctor role.
      parameters:
      - description: Create diagnosis payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.swaggerCreateDiagnosisRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerDiagnosisResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a diagnosis
      tags:
      - Diagnoses
  /diagnoses/{id}:
    get:
      description: Returns a single diagnosis by ID. Requires doctor or patient role.
      parameters:
      - description: Diagnosis ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerDiagnosisResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a diagnosis
      tags:
      - Diagnoses
  /diagnoses/{id}/dismiss:
    patch:
      description: Marks a diagnosis as dismissed. Requires doctor role.
      parameters:
      - description: Diagnosis ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.swaggerDiagnosisResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Dismiss a diagnosis
      tags:
      - Diagnoses
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Example: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"